+++ tests/testthat/test-prepInputs.R
@@ -278,7 +278,7 @@ test_that("interactive prepInputs", {
                            targetFile = paste0(x, "_currmean.asc"),
                            archive = paste0(x, "_current.zip"),
                            url = url,
-                           alsoExtract = NA,
+                           alsoExtract = NULL,
                            destinationPath = tmpdir,
                            overwrite = TRUE
                          )
@@ -401,7 +401,7 @@ test_that("preProcess doesn't work", {
       )
     )
   )
-  runTest("1_2_5_6_8", vectorType(), 5, mess, expectedMess = expectedMessage,
+  runTest("1_2_5_6_8_9", vectorType(), 5, mess, expectedMess = expectedMessage,
           filePattern = "Shapefile", tmpdir = tmpdir, test = test)
   unlink(dir(tmpdir, full.names = TRUE))
 
@@ -560,7 +560,7 @@ test_that("preProcess doesn't work", {
       )
     )
   )
-  runTest("1_2_5_6_8", vectorType(), 5, mess, expectedMess = expectedMessage,
+  runTest("1_2_5_6_8_9", vectorType(), 5, mess, expectedMess = expectedMessage,
           filePattern = "Shapefile", tmpdir = tmpdir, test = test)
   unlink(dir(tmpdir, full.names = TRUE))
 
@@ -695,7 +695,7 @@ test_that("preProcess doesn't work", {
       )
     )
   )
-  runTest("1_2_5_6_8", vectorType(), 5, mess, expectedMess = expectedMessage,
+  runTest("1_2_5_6_8_9", vectorType(), 5, mess, expectedMess = expectedMessage,
           filePattern = "Shapefile", tmpdir = tmpdir, test = test)
   unlink(dir(tmpdir, full.names = TRUE))
 
@@ -1398,7 +1398,7 @@ test_that("lightweight tests 2 for code coverage", {
 test_that("options inputPaths", {
   skip_on_cran()
   if (!requireNamespace("geodata", quietly = TRUE)) skip("Need geodata package")
-
+  if (getRversion() <= "4.1.3") skip("geodata::gadm seems to time out on R <= 4.1.3")
   testInitOut <- testInit(c("terra", "geodata"),
                           opts = list("reproducible.inputPaths" = NULL,
                                       "reproducible.inputPathsRecursive" = FALSE))
@@ -1475,7 +1475,7 @@ test_that("options inputPaths", {
                           getDataFn = dlFun1)
     )
   )
-  expect_true(sum(grepl(paste0("Hardlinked", ".*: "), mess1)) == 1)
+  expect_true(sum(grepl(paste0("Hardlinked", ".*:"), mess1)) == 1)
 
   # Now two folders - file not in destinationPath, not in 1st inputPaths, but yes 2nd
   #   should hardlink from 2nd IP to destinationPath, make sure CHECKSUMS.txt is correct in both
@@ -1494,7 +1494,7 @@ test_that("options inputPaths", {
                           destinationPath = tmpdir3)
     )
   )
-  expect_true(sum(grepl(paste0(hardlinkMessagePrefixForGrep, ": ", tmpdir3), mess1)) == 1)
+  expect_true(sum(grepl(paste0(hardlinkMessagePrefixForGrep, ":\n", tmpdir3), mess1)) == 1)
 
 
   # THIS NEXT ONE DOESN"T PASS ON GA on WINDOWS, skip it
@@ -1517,8 +1517,8 @@ test_that("options inputPaths", {
                             destinationPath = tmpdir1)
       )
     )
-    expect_true(sum(grepl(paste0(hardlinkMessagePrefixForGrep, ": ", file.path(tmpdir1, theFile)), mess1)) == 1)
-    expect_true(sum(grepl(paste0("",whPointsToMessForGrep," ", file.path(tmpdir3, theFile)), mess1)) == 1)
+    expect_true(sum(grepl(paste0(hardlinkMessagePrefixForGrep, ":\n", file.path(tmpdir1, theFile)), mess1)) == 1)
+    expect_true(sum(grepl(paste0("",whPointsToMessForGrep,"\n", file.path(tmpdir3, theFile)), mess1)) == 1)
     expect_true(sum(basename(dir(file.path(tmpdir), recursive = TRUE)) %in% theFile) == 2)
 
   }